Move region logic out of the view
