+++ cubedash/_pages.py
@@ -42,7 +42,6 @@ app.register_blueprint(_reports.bp)
 _LOG = structlog.getLogger()
 
 _HARD_SEARCH_LIMIT = 500
-_WRS_PATH_ROW = Path(__file__).parent / 'data' / 'WRS2_descending' / 'WRS2_descending.shp'
 
 
 # @app.route('/')
@@ -56,152 +55,23 @@ def overview_page(product_name: str = None,
                   day: int = None):
     product, product_summary, selected_summary = _load_product(product_name, year, month, day)
 
-    datasets = None
-    regions = None
-
-    footprint_wrs84 = None
-
-    if selected_summary and selected_summary.dataset_count:
-        if selected_summary.footprint_geometry:
-            start = time.time()
-            from_crs = pyproj.Proj(init=selected_summary.footprint_crs)
-            to_crs = pyproj.Proj(init='epsg:4326')
-            footprint_wrs84 = transform(
-                lambda x, y: pyproj.transform(from_crs, to_crs, x, y),
-                selected_summary.footprint_geometry
-            )
-            _LOG.info(
-                'overview.footprint_size_diff',
-                from_len=len(selected_summary.footprint_geometry.wkt),
-                to_len=len(footprint_wrs84.wkt),
-            )
-            _LOG.debug('overview.footprint_proj', time_sec=time.time() - start)
-
-        # The per-dataset view is less useful now that we show grids separately.
-        # datasets = None if selected_summary.dataset_count > 1000
-        # else get_datasets_geojson(product_name, year, month, day)
-
-        start = time.time()
-        regions = get_region_counts(
-            selected_summary.region_dataset_counts,
-            footprint_wrs84,
-            product
-        ) if selected_summary.region_dataset_counts else None
-
-        _LOG.debug('overview.region_gen', time_sec=time.time() - start)
-
     return flask.render_template(
         'overview.html',
         year=year,
         month=month,
         day=day,
 
-        regions_geojson=regions,
-        datasets_geojson=datasets,
+        # Which data to preload with the page?
+        regions_geojson=None,  # _model.get_regions_geojson(product_name, year, month, day),
+        datasets_geojson=None,  # _model.get_datasets_geojson(product_name, year, month, day),
+        footprint_wrs84=_model.get_footprint_geojson(product_name, year, month, day),
+
         product=product,
         # Summary for the whole product
         product_summary=product_summary,
         # Summary for the users' currently selected filters.
         selected_summary=selected_summary,
-        footprint_wrs84=footprint_wrs84,
     )
-
-
-def get_region_counts(
-        region_counts: Counter[str],
-        footprint: MultiPolygon,
-        product: DatasetType
-) -> Optional[Dict]:
-    region_geometry = _region_geometry_function(product, footprint)
-    if not region_geometry:
-        return None
-
-    low, high = min(region_counts.values()), max(region_counts.values())
-    return {
-        'type': 'FeatureCollection',
-        'properties': {
-            'region_item_name': 'Tile',
-            'min_count': low,
-            'max_count': high,
-        },
-        'features': [
-            {
-                'type': 'Feature',
-                'geometry': region_geometry(region_code).__geo_interface__,
-                'properties': {
-                    'region_code': region_code,
-                    'count': region_counts[region_code]
-                }
-            } for region_code in region_counts
-        ]
-    }
-
-
-def _from_xy_region_code(region_code: str):
-    x, y = region_code.split('_')
-    return int(x), int(y)
-
-
-def _region_geometry_function(product, footprint):
-    grid_spec = product.grid_spec
-    md_fields = product.metadata_type.dataset_fields
-    # TODO: Geometry for other types of regions (path/row, MGRS)
-
-    # hltc has a grid spec, but most attributes are missing, so grid_spec functions fail.
-    # Therefore: only assume there's a grid if tile_size is specified. TODO: Is the product wrong?
-    if grid_spec and grid_spec.tile_size:
-
-        def region_geometry(region_code: str) -> shapely.geometry.GeometryCollection:
-            """
-            Get a whole polygon for a gridcell
-            """
-            extent = grid_spec.tile_geobox(_from_xy_region_code(region_code)).geographic_extent
-            # TODO: The ODC Geometry __geo_interface__ breaks for some products
-            # (eg, when the inner type is a GeometryCollection?)
-            # So we're now converting to shapely to do it.
-            # TODO: Is there a nicer way to do this?
-            # pylint: disable=protected-access
-            shapely_extent = shapely.wkb.loads(extent._geom.ExportToWkb())
-
-            return shapely_extent
-    elif 'sat_path' in md_fields:
-        path_row_shapes = _get_path_row_shapes()
-
-        def region_geometry(region_code: str) -> shapely.geometry.GeometryCollection:
-            return path_row_shapes[_from_xy_region_code(region_code)]
-    else:
-        _LOG.info('region.geom.unknown', product_name=product.name)
-        return None
-
-    if footprint is None:
-        return region_geometry
-    else:
-        footprint_boundary = shapely.prepared.prep(footprint.boundary)
-
-        def region_geometry_cut(region_code: str) -> shapely.geometry.GeometryCollection:
-            """
-            Cut the polygon down to the footprint
-            """
-            shapely_extent = region_geometry(region_code)
-
-            # We only need to cut up tiles that touch the edges of the footprint (including inner "holes")
-            # Checking the boundary is ~2.5x faster than running intersection() blindly, from my tests.
-            if footprint_boundary.intersects(shapely_extent):
-                return footprint.intersection(shapely_extent)
-            else:
-                return shapely_extent
-
-        return region_geometry_cut
-
-
-@functools.lru_cache()
-def _get_path_row_shapes():
-    path_row_shapes = {}
-    with fiona.open(str(_WRS_PATH_ROW)) as f:
-        for k, item in f.items():
-            prop = item['properties']
-            path_row_shapes[prop['PATH'], prop['ROW']] = shape(item['geometry'])
-    return path_row_shapes
 
 
 # @app.route('/datasets')
